make compile_to_module_outputs a method
As a side effect, Output.cpp goes away, and Output.h -> Outputs.h to
reflect its only content
