+++ src/Module.cpp
@@ -1,9 +1,19 @@
 #include "Module.h"
+
+#include <fstream>
+
+#include "CodeGen_C.h"
 #include "Debug.h"
+#include "LLVM_Headers.h"
+#include "LLVM_Output.h"
+#include "LLVM_Runtime_Linker.h"
+#include "Outputs.h"
+#include "StmtToHtml.h"
 
 namespace Halide {
 
 namespace Internal {
+
 struct ModuleContents {
     mutable RefCount ref_count;
     std::string name;
@@ -21,7 +31,8 @@ template<>
 EXPORT void destroy<ModuleContents>(const ModuleContents *f) {
     delete f;
 }
-}
+
+}  // namespace Internal
 
 Module::Module(const std::string &name, const Target &target) :
     contents(new Internal::ModuleContents) {
@@ -77,6 +88,62 @@ Module link_modules(const std::string &name, const std::vector<Module> &modules)
     }
 
     return output;
+}
+
+void Module::compile(const Outputs &output_files) const {
+    if (!output_files.object_name.empty() || !output_files.assembly_name.empty() ||
+        !output_files.bitcode_name.empty() || !output_files.llvm_assembly_name.empty()) {
+        llvm::LLVMContext context;
+        std::unique_ptr<llvm::Module> llvm_module(compile_module_to_llvm_module(*this, context));
+
+        if (!output_files.object_name.empty()) {
+            if (target().arch == Target::PNaCl) {
+                compile_llvm_module_to_llvm_bitcode(*llvm_module, output_files.object_name);
+            } else {
+                compile_llvm_module_to_object(*llvm_module, output_files.object_name);
+            }
+        }
+        if (!output_files.assembly_name.empty()) {
+            if (target().arch == Target::PNaCl) {
+                compile_llvm_module_to_llvm_assembly(*llvm_module, output_files.assembly_name);
+            } else {
+                compile_llvm_module_to_assembly(*llvm_module, output_files.assembly_name);
+            }
+        }
+        if (!output_files.bitcode_name.empty()) {
+            compile_llvm_module_to_llvm_bitcode(*llvm_module, output_files.bitcode_name);
+        }
+        if (!output_files.llvm_assembly_name.empty()) {
+            compile_llvm_module_to_llvm_assembly(*llvm_module, output_files.llvm_assembly_name);
+        }
+    }
+    if (!output_files.c_header_name.empty()) {
+        std::ofstream file(output_files.c_header_name.c_str());
+        Internal::CodeGen_C cg(file,
+                               target().has_feature(Target::CPlusPlusMangling) ?
+                               Internal::CodeGen_C::CPlusPlusHeader : Internal::CodeGen_C::CHeader,
+                               output_files.c_header_name);
+        cg.compile(*this);
+    }
+    if (!output_files.c_source_name.empty()) {
+        std::ofstream file(output_files.c_source_name.c_str());
+        Internal::CodeGen_C cg(file,
+                               target().has_feature(Target::CPlusPlusMangling) ?
+                               Internal::CodeGen_C::CPlusPlusImplementation : Internal::CodeGen_C::CImplementation);
+        cg.compile(*this);
+    }
+    if (!output_files.stmt_name.empty()) {
+        std::ofstream file(output_files.stmt_name.c_str());
+        file << *this;
+    }
+    if (!output_files.stmt_html_name.empty()) {
+        Internal::print_to_html(output_files.stmt_html_name, *this);
+    }
+}
+
+void compile_standalone_runtime(std::string object_filename, Target t) {
+    Module empty("standalone_runtime", t.without_feature(Target::NoRuntime).without_feature(Target::JIT));
+    empty.compile(Outputs().object(object_filename));
 }
 
 }
